liquid_time-constant_networks/requirements.txt
numpy 
scipy
pandas
matplotlib
scikit-learn

# ---実験管理・設定---
wandb
omegaconf
hydra-core

# ---PyTorch関連---
torch
torchvision
pytorch-lightning

# models
transformers

# UI/表示
rich

# 追加依存関係
einops
opt_einsum
ptflops

ncps

# ---TensorFlow関連---
tensorflow==2.14.0
tensorboard==2.14.1
protobuf==3.20.3

fvcore==0.1.5